Breaking: Take a struct as argument in callbacks
This makes it a lot simpler to write the callbacks since
the sole parameter will have a name that can be guessed
instead of having to be looked up. It will also make it
easier to ignore unused parameters and make changes less
breaking so I think the extra effort of having to unpack
the struct will be well worth it.